backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.7.0
Pillow==11.1.0
python-dotenv==1.1.1
requests==2.31.0
